+++ semcore/base-components/src/components/popper/Popper.tsx
@@ -606,16 +606,7 @@ function Trigger(props: PopperTriggerProps & IRootComponentProps & InnerPopperTr
   );
 
   React.useEffect(() => {
-    if (!triggerRef.current) return;
-
-    const originFocus = triggerRef.current.focus;
-
-    triggerRef.current.focus = function () {
-      lastInteraction.internalSetter('keyboard');
-      originFocus.apply(this);
-    };
-
-    triggerRef.current.addEventListener('focusin', handleFocus);
+    triggerRef.current?.addEventListener('focusin', handleFocus);
 
     return () => {
       triggerRef.current?.removeEventListener('focusin', handleFocus);