+++ semcore/core/CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Added
 
-- `internalSetter` method to LastInteractionType utility class for manually set `keyboard` for programmatic  `.focus()` calls.
+- Logic to manually set `keyboard` for type in LastInteractionType utility class for programmatic `.focus()` calls.
 
 ## [16.0.0] - 2025-05-19